Extract preprocessors into dedicated tracemill.preprocessors package
Move all 5 preprocessors (openhands, goose, cline, pydantic_ai, smolagents)
out of the god module mapped_json.py into their own package with:
- registry.py: PreprocessorFn type, register_preprocessor decorator, get_preprocessor lookup
- One module per framework preprocessor
- __init__.py handles registration via imports

mapped_json.py now imports get_preprocessor from the package instead of
maintaining an inline registry + all function definitions.
